Fix extra whitespace flake8 error
Add test for adr query --debug

Added a test for when adr query --debug is run. This has simply
been added to the original query test and is measured by asserting
if the debug_url returned by the format_query() function with mock
data input is either None (for no --debug) or is equal to the
expected debug_url (for --debug included).

Currently, this is tested with a separate file triggered by the 'debug'
key set to either True or False. I originally experimented with just
the 'saved_as' key triggering debug or no debug in the test, but that
was not achievable as 'saved_as' is used in query.py to set the
build_debug_url() in motion.

Updated test for adr query --debug
Amended test to check both debug_url and result in both 'debug: True'
and 'debug: False'. Consolidated .test file to include mock and expected
data for both scenarios in one file. Removed unnecessary 'es_query' and
'timing' keys from data. Updated debug_url to 'https://'.