Remove seta_accuracy.test
This started failing because the test didn't pass in --from/--to as
arguments (and the script uses those in the post-processing).

As a result the expected data got mismatched from the actual data.

Unfortunately the 'adr seta_accuracy' command itself stopped working in
the meantime due to the ActiveData migration. So since both test and
recipe are broken, I'm going to remove the test. Once the recipe is
fixed, we can re-generate the test with a fresh set of mocked data.