Update to the right time counting
That was my mistake, commited wrong one out of compared 2 methods in Python's time module. time.clock() is meant to return processor time, so that any time spent in time.sleep() is not counted on *nix systems, but counted on Windows. time.time() returns time in seconds since epoch at the moment of call on all systems.